LayoutTests/platform/mac-mojave/fast/text/international/system-language/navigator-language/navigator-language-ru-expected.txt

FAIL navigator.language should be ru. Was ru-ru.
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
